requirements.txt
numpy
pandas
pyxlsb
sqlalchemy
xmltodict
python-dotenv
requests
keras
opencv-python
tensorflow
pillow